Revert "add customs hooks"
