+++ src/components/MovieDetails.jsx
@@ -1,12 +1,10 @@
-import { useState, useEffect, useRef } from "react";
+import { useState, useEffect } from "react";
 import { FaArrowLeft } from "react-icons/fa";
 import { toast } from "react-toastify";
 import StarRating from "./StarRating";
 import Loader from "./Loader";
-import { useKey } from "../hooks/useKey";
 
 const KEY = "59350f2c";
-
 export default function MovieDetails({
   selelectedId,
   onCloseMovie,
@@ -16,11 +14,6 @@ export default function MovieDetails({
   const [movie, setMovie] = useState({});
   const [isLoading, setIsLoading] = useState(false);
   const [userRating, setUserRating] = useState("");
-  const countRef = useRef(0);
-
-  useEffect(() => {
-    if (userRating) countRef.current++;
-  }, [userRating]);
 
   const isWatched = watched.map((movie) => movie.imdbID).includes(selelectedId);
   const watchedUserRating = watched.find(
@@ -49,7 +42,6 @@ export default function MovieDetails({
       imdbRating: Number(imdbRating),
       runtime: Number(runtime.split(" ").at(0)),
       userRating,
-      countRatingDecisions: countRef.current,
     };
 
     onAddWatched(newWatchedMovie);
@@ -57,7 +49,18 @@ export default function MovieDetails({
     toast.success("Film ajouté !");
   }
 
-  useKey("Escape", onCloseMovie);
+  useEffect(() => {
+    function callback(e) {
+      if (e.code === "Escape") {
+        onCloseMovie();
+      }
+    }
+    document.addEventListener("keydown", callback);
+
+    return function () {
+      document.removeEventListener("keydown", callback);
+    };
+  }, [onCloseMovie]);
 
   useEffect(() => {
     async function getMoviesDetails() {
@@ -79,7 +82,7 @@ export default function MovieDetails({
     document.title = `Film | ${title}`;
 
     return function () {
-      document.title = "CornFlix";
+      document.title = "usePopcorn";
     };
   }, [title]);
 
@@ -106,6 +109,7 @@ export default function MovieDetails({
               </p>
             </div>
           </header>
+
           <section>
             <div className="rating">
               {!isWatched ? (
