+++ src/components/Search.jsx
@@ -1,25 +1,11 @@
-import { useRef } from "react";
-import { useKey } from "../hooks/useKey";
-
 export default function Search({ query, setQuery }) {
-  const inputEl = useRef(null);
-
-  useKey("Enter", function () {
-    if (document.activeElement === inputEl.current) {
-      return;
-    }
-    inputEl.current.focus();
-    setQuery("");
-  });
-
   return (
     <input
       className="search"
       type="text"
-      placeholder="Rechercher des films"
+      placeholder="Rechercher des films..."
       value={query}
       onChange={(e) => setQuery(e.target.value)}
-      ref={inputEl}
     />
   );
 }
