+++ src/components/WatchSummary.jsx
@@ -2,11 +2,6 @@ const average = (arr) =>
   arr.reduce((acc, cur, i, arr) => acc + cur / arr.length, 0);
 
 export default function WatchSummary({ watched }) {
-  if (!Array.isArray(watched)) {
-    console.error("Erreur : `watched` n'est pas un tableau", watched);
-    return null; // Ou un fallback UI
-  }
-
   const avgImdbRating = average(watched.map((movie) => movie.imdbRating));
   const avgUserRating = average(watched.map((movie) => movie.userRating));
   const avgRuntime = average(watched.map((movie) => movie.runtime));
@@ -29,7 +24,7 @@ export default function WatchSummary({ watched }) {
         </p>
         <p>
           <span>⏳</span>
-          <span>{avgRuntime.toFixed(2)} min</span>
+          <span>{avgRuntime} min</span>
         </p>
       </div>
     </div>
