--- src/hooks/useKey.jsx
@@ -1,17 +1,0 @@
-import { useEffect } from "react";
-
-export function useKey(key, action) {
-  useEffect(() => {
-    function callback(e) {
-      if (e.code.toLowerCase() === key.toLowerCase()) {
-        action();
-      }
-    }
-
-    document.addEventListener("keydown", callback);
-
-    return () => {
-      document.removeEventListener("keydown", callback);
-    };
-  }, [key, action]);
-}
