--- src/hooks/useMovies.jsx
@@ -1,64 +1,0 @@
-import { useState, useEffect } from "react";
-
-const KEY = process.env.REACT_APP_API_KEY;
-
-export function useMovies(query) {
-  const [movies, setMovies] = useState([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState("");
-
-  useEffect(
-    function () {
-      // callback?.();
-
-      const controller = new AbortController();
-
-      async function fetchMovies() {
-        try {
-          setIsLoading(true);
-          setError("");
-
-          const res = await fetch(
-            `http://www.omdbapi.com/?i=tt3896198&apikey=${KEY}&s=${query}`,
-            { signal: controller.signal }
-          );
-
-          if (!res.ok)
-            throw new Error(
-              "Une erreur est survenue lors de la récupération des films"
-            );
-
-          const data = await res.json();
-
-          if (data.Response === "False") throw new Error("Film non trouvé");
-
-          setMovies(data.Search);
-          setError("");
-        } catch (err) {
-          if (err.name !== "AbortError") {
-            console.log(err.message);
-            setError(err.message);
-          }
-        } finally {
-          setIsLoading(false);
-        }
-      }
-
-      if (query.length < 3) {
-        setMovies([]);
-        setError("");
-        return;
-      }
-
-      // handleCloseMovie();
-      fetchMovies();
-
-      return function () {
-        controller.abort();
-      };
-    },
-    [query]
-  );
-
-  return { movies, isLoading, error };
-}
